requirements.txt
pandas
numpy
pyyaml
fastapi
uvicorn
pytest
scikit-learn
scipy
